fixing negative event numbers. Details: https://its.cern.ch/jira/browse/ATLASVPONE-287. (VP1Gui-00-08-11)
    * fixing negative event numbers.
	  Details: https://its.cern.ch/jira/browse/ATLASVPONE-287
	* new tag: VP1Gui-00-08-11


Former-commit-id: d3fad031dfddc1ced3b82cfb77ee621134ebce92

+++ graphics/VP1/VP1Gui/VP1Gui/VP1EventDisplayEditor.h
@@ -37,7 +37,7 @@ class VP1EventDisplayEditor : public QWidget {
 
 public:
 
-	VP1EventDisplayEditor(VP1MainWindow* mainWindow, QList<int> listRunEventNumberTimestamp);
+	VP1EventDisplayEditor(VP1MainWindow* mainWindow, QList<unsigned long long> listRunEventNumberTimestamp);
 	virtual ~VP1EventDisplayEditor();
 
 //	QGraphicsView* getView();
@@ -83,7 +83,7 @@ private:
 	VP1MainWindow* _mainWindow;
 
 	int _runNumber;
-	int _eventNumber;
+	unsigned long long _eventNumber;
 	unsigned _eventTimestamp;
 
 //	QSplitter *h1Splitter;
